+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 #### v1.14.0 `2017-03-??`
 - `Added`
     - With the `--cli-framework` option a CLI project can be generated. Done by [@raphaelstolt](https://github.com/raphaelstolt). See [#175](https://github.com/jonathantorres/construct/issues/175).
+- `Fixed`
+    - The package `vlucas/phpdotenv` is added as a non development requirement. Done by [@raphaelstolt](https://github.com/raphaelstolt). See [#176](https://github.com/jonathantorres/construct/issues/176).
 
 #### v1.13.1 `2017-03-01`
 - `Fixed`